Revert part of 277d3f3af4e58f5f0daa53dde661aa3a5624d9e4
The fix for issue #119 would be to set offset to 0 or do not pass it anyways.
@serjmac will lose the first record in the list on every query.
That's why hasPrevPage is correctly set to true.

I rely on the existing behaviour because if a user changes the number of elements per page (limit e.g. 5 > 10) and was already on page 2 of the previous query (e.g. offset 5) he needs to be able to go back to the complete page 1 starting at record 0.